Remove unnecessary validate call, clean up dependencies
