+++ src/selectors/navigation.js
@@ -9,86 +9,6 @@ import * as sections from 'constants/sections'
 import { hideSelectiveService } from 'validators/selectiveservice'
 import { hideDisciplinaryProcedures } from 'validators/militarydisciplinary'
 import { hideExistingConditions } from 'validators/psychological'
-import { formIsLocked } from 'validators'
-
-import { formHasErrors, reduceSubsections } from 'helpers/navigation'
-
-const getSectionErrors = (state, props) => {
-  const { application } = state
-  const { Errors } = application
-
-  const { topSection, section } = props
-
-  const sectionErrors = topSection
-    ? Errors[topSection]
-    : Errors[section.name]
-
-  if (!sectionErrors) { return [] }
-
-  if (!section.subsections) {
-    return sectionErrors.filter(e => e.subsection === section.name)
-  }
-
-  const flatSections = reduceSubsections(section.subsections)
-
-  return sectionErrors.filter(s => (
-    flatSections.find(i => i.name === s.subsection)
-  ))
-}
-
-const getSectionCompleted = (state, props) => {
-  const { form } = state
-  const { section } = props
-
-  if (section.subsections) {
-    // Check complete status of each subsection
-    const flatSections = reduceSubsections(section.subsections)
-    const sectionsWithData = flatSections.filter(s => !!s.storeKey)
-    return sectionsWithData.length > 0 && sectionsWithData.every((s) => {
-      const sectionData = form && form[s.key]
-      return sectionData && sectionData.complete === true
-    })
-  }
-
-  const sectionData = form && form[section.key]
-  if (!sectionData) return false
-  return sectionData.complete === true
-}
-
-const getSectionLocked = (state, props) => {
-  const { section } = props
-  const formLocked = formIsLocked(state.application)
-
-  // Special cases
-  switch (section.key) {
-    case sections.REVIEW_AND_SUBMIT_SUBMIT: {
-      return formLocked || formHasErrors(state)
-    }
-
-    case sections.REVIEW_AND_SUBMIT_PRINT:
-      return !formLocked
-
-    default:
-      return formLocked
-  }
-}
-
-export const sectionHasErrorsSelector = createSelector(
-  getSectionErrors,
-  (errors = []) => ({
-    errors: errors.some(e => e.valid === false),
-  })
-)
-
-export const sectionIsValidSelector = createSelector(
-  getSectionCompleted,
-  completed => ({ completed })
-)
-
-export const sectionIsLockedSelector = createSelector(
-  getSectionLocked,
-  locked => ({ locked })
-)
 
 export const hideSelectiveServiceSelector = (state) => {
   const { application } = state
